% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{PheNorm}
\alias{PheNorm}
\title{Fit the phenotyping algorithm with EHR features. The function requires a surrogate (ICD) and
the health utilization as its input and can leverage other EHR features (optional) to assist
risk prediction. Uses the healthcare utilization in input.}
\usage{
PheNorm(
  nm.logS.ori,
  nm.utl,
  dat,
  nm.X = NULL,
  corrupt.rate = 0.3,
  train.size = 1e+05
)
}
\arguments{
\item{nm.logS.ori}{name of the surrogates (log(ICD+1), log(NLP+1) and log(ICD+NLP+1)}

\item{nm.utl}{name of healthcare utilization (e.g. note count, encounter_num etc)}

\item{dat}{all data columns need to be log-transformed and need column names}

\item{nm.X}{additional features other than the main ICD and NLP}

\item{corrupt.rate}{rate for random corruption denoising, between 0 and 1}

\item{train.size}{size of training sample, default value 10000}
}
\value{
S.norm
}
\description{
Fit the phenotyping algorithm with EHR features. The function requires a surrogate (ICD) and
the health utilization as its input and can leverage other EHR features (optional) to assist
risk prediction. Uses the healthcare utilization in input.
}
