% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{PheNorm.Prob_noUTL}
\alias{PheNorm.Prob_noUTL}
\title{Fit the phenotyping algorithm with EHR features. The function requires a surrogate (ICD) and
the health utilization as its input and can leverage other EHR features (optional) to assist
risk prediction.}
\usage{
PheNorm.Prob_noUTL(
  nm.logS.ori,
  dat,
  nm.X = NULL,
  corrupt.rate = 0.3,
  train.size = 10000
)
}
\arguments{
\item{nm.logS.ori}{name of the surrogates (log(ICD+1), log(NLP+1) and log(ICD+NLP+1)}

\item{dat}{all data columns need to be log-transformed and need column names}

\item{nm.X}{additional features other than the main ICD and NLP}

\item{corrupt.rate}{rate for random corruption denoising, between 0 and 1}

\item{train.size}{size of training sample, default value 10000}
}
\value{
list containing probability and beta coefficient
}
\description{
Fit the phenotyping algorithm with EHR features. The function requires a surrogate (ICD) and
the health utilization as its input and can leverage other EHR features (optional) to assist
risk prediction.
}
